chore(bridge):change test file name
