+++ src/test/java/bridge/RunMainTest.java
@@ -10,8 +10,7 @@ import camp.nextstep.edu.missionutils.test.NsTest;
 import java.util.List;
 import org.junit.jupiter.api.Test;
 
-class ApplicationTest extends NsTest {
-
+public class RunMainTest extends NsTest {
     private static final String ERROR_MESSAGE = "[ERROR]";
 
     @Test
@@ -48,13 +47,49 @@ class ApplicationTest extends NsTest {
         });
     }
 
+    @Test
+    void 기능_실패_테스트() {
+        assertRandomNumberInRangeTest(() -> {
+            run("3", "U", "U", "R", "U", "D", "U");
+            assertThat(output()).contains(
+                    "[ O | X ]",
+                    "[   |   ]",
+                    "최종 게임 결과",
+                    "[ O |   | O ]",
+                    "[   | O |   ]",
+                    "게임 성공 여부: 성공",
+                    "총 시도한 횟수: 2"
+            );
+
+            int upSideIndex = output().indexOf("[ O |   | O ]");
+            int downSideIndex = output().indexOf("[   | O |   ]");
+            assertThat(upSideIndex).isLessThan(downSideIndex);
+        }, 1, 0, 1);
+    }
+
+    @Test
+    void 기능_종료_테스트() {
+        assertRandomNumberInRangeTest(() -> {
+            run("3", "U", "U", "Q");
+            assertThat(output()).contains(
+                    "[ O | X ]",
+                    "[   |   ]",
+                    "최종 게임 결과",
+                    "[ O | X ]",
+                    "[   |   ]",
+                    "게임 성공 여부: 실패",
+                    "총 시도한 횟수: 1"
+            );
+
+        }, 1, 0, 1);
+    }
+
     @Override
     protected void runMain() {
         Application.main(new String[]{});
     }
 
     static class TestNumberGenerator implements BridgeNumberGenerator {
-
         private final List<Integer> numbers;
 
         TestNumberGenerator(List<Integer> numbers) {
